PEEK-541 bug: CompilerQueues run out of Integer sequence space.
The sequences are and all relevent columns will be updated to BigInteger
fields.

+++ peek_plugin_diagram/_private/storage/DispIndex.py
@@ -1,10 +1,9 @@
-from sqlalchemy import Column, Integer, Index
-from vortex.Tuple import addTupleType, Tuple
-
 from peek_abstract_chunked_index.private.tuples.ACIProcessorQueueTupleABC import \
     ACIProcessorQueueTupleABC
 from peek_plugin_diagram._private.PluginNames import diagramTuplePrefix
 from peek_plugin_diagram._private.storage.DeclarativeBase import DeclarativeBase
+from sqlalchemy import Column, Index, BigInteger
+from vortex.Tuple import addTupleType, Tuple
 
 
 @addTupleType
@@ -13,8 +12,8 @@ class DispIndexerQueue(Tuple, DeclarativeBase,
     __tablename__ = 'DispCompilerQueue'
     __tupleType__ = diagramTuplePrefix + __tablename__
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    dispId = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    dispId = Column(BigInteger, primary_key=True)
 
     __table_args__ = (
         Index("idx_DispCompQueue_dispId", dispId, unique=False),
